+++ src/windows/settings/General.tsx
@@ -12,6 +12,9 @@ import {
 } from '~/components/ui/select';
 import { enable, isEnabled, disable } from 'tauri-plugin-autostart-api';
 import { CheckedState } from '@radix-ui/react-checkbox';
+import { useSetting } from '~/store/settings';
+import { Slider } from '~/components/ui/slider';
+import { Separator } from '~/components/ui/separator';
 
 // placeholders for future features
 const showPlaceholders = false;
@@ -20,7 +23,134 @@ function Description({ children }: { children: ReactNode }) {
   return <p className="text-sm mb-2">{children}</p>;
 }
 
+function LayoutGrid({ children }: { children: ReactNode }) {
+  return (
+    <div className="grid grid-cols-[160px_1fr] items-center gap-y-2 gap-x-4 auto-rows-fr">
+      {children}
+    </div>
+  );
+}
+
 export function General() {
+  const [sounds, setSounds] = useSetting<boolean>('sound-effects', true);
+  const [volume, setVolume] = useSetting<number>('sound-volume', 1);
+
+  function handleToggleSounds(value: CheckedState) {
+    setSounds(value === true);
+  }
+
+  function handleVolumeChange(value: number[]) {
+    setVolume(value[0]);
+  }
+
+  return (
+    <div className="flex flex-col gap-8">
+      <section>
+        <LayoutGrid>
+          <div className="text-sm justify-self-end items-center">Startup:</div>
+          <StartupSetting />
+        </LayoutGrid>
+      </section>
+      <Separator />
+      <section>
+        <div className="flex flex-col gap-5">
+          <LayoutGrid>
+            <>
+              <Label className="col-start-2 flex items-center justify-self-start">
+                <Checkbox
+                  className="mr-2"
+                  checked={sounds}
+                  onCheckedChange={handleToggleSounds}
+                />
+                <div className="text-sm justify-self-end">Sound effects</div>
+              </Label>
+            </>
+            <>
+              <div className="text-sm justify-self-end">Volume:</div>
+              <Slider
+                className="max-w-[200px]"
+                defaultValue={[volume]}
+                onValueCommit={handleVolumeChange}
+                disabled={!sounds}
+                min={0}
+                max={1}
+                step={0.01}
+              />
+            </>
+            <>
+              <SoundSelect
+                label="Start recording:"
+                soundEvent="sound-start"
+                disabled={!sounds}
+              />
+            </>
+            <>
+              <SoundSelect
+                label="Stop recording:"
+                soundEvent="sound-stop"
+                disabled={!sounds}
+              />
+            </>
+            <>
+              <SoundSelect
+                label="Transcription complete:"
+                soundEvent="sound-complete"
+                disabled={!sounds}
+              />
+            </>
+            <>
+              <div className="text-xs col-start-2">
+                More sounds coming soon!
+              </div>
+            </>
+          </LayoutGrid>
+        </div>
+      </section>
+      {showPlaceholders && (
+        <section>
+          <SettingTitle>Overlay</SettingTitle>
+          <div className="flex items-center">
+            <Label className="flex items-center">
+              <Checkbox className="mr-2" />
+              <span>Overlay Position</span>
+            </Label>
+          </div>
+        </section>
+      )}
+      {showPlaceholders && (
+        <section>
+          <SettingTitle>Prompt</SettingTitle>
+          <Description>
+            The prompt to provide the model. This is typically used to correct
+            spelling. Limit of 250 characters.
+          </Description>
+          <div className="flex items-center">
+            <Textarea className="max-w-lg h-[100px] resize-none" />
+          </div>
+        </section>
+      )}
+      {showPlaceholders && (
+        <section>
+          <SettingTitle>Audio Device</SettingTitle>
+          <Description>
+            The prompt to provide the model. This is typically used to correct
+            spelling. Limit of 250 characters.
+          </Description>
+          <Select>
+            <SelectTrigger className="w-[180px]">
+              <SelectValue placeholder="Theme" />
+            </SelectTrigger>
+            <SelectContent>
+              <SelectItem value="light">Default</SelectItem>
+            </SelectContent>
+          </Select>
+        </section>
+      )}
+    </div>
+  );
+}
+
+function StartupSetting() {
   const [startupEnabled, setStartupEnabled] = useState(false);
 
   useEffect(() => {
@@ -41,60 +171,57 @@ export function General() {
   }
 
   return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <SettingTitle>Startup</SettingTitle>
-        <div className="flex items-center">
-          <Label className="flex items-center">
-            <Checkbox
-              className="mr-2"
-              checked={startupEnabled}
-              onCheckedChange={handleStartupEnabledChange}
-            />
-            <span>Start at login</span>
-          </Label>
-        </div>
-      </div>
-      {showPlaceholders && (
-        <div>
-          <SettingTitle>Overlay</SettingTitle>
-          <div className="flex items-center">
-            <Label className="flex items-center">
-              <Checkbox className="mr-2" />
-              <span>Overlay Position</span>
-            </Label>
-          </div>
-        </div>
-      )}
-      {showPlaceholders && (
-        <div>
-          <SettingTitle>Prompt</SettingTitle>
-          <Description>
-            The prompt to provide the model. This is typically used to correct
-            spelling. Limit of 250 characters.
-          </Description>
-          <div className="flex items-center">
-            <Textarea className="max-w-lg h-[100px] resize-none" />
-          </div>
-        </div>
-      )}
-      {showPlaceholders && (
-        <div>
-          <SettingTitle>Audio Device</SettingTitle>
-          <Description>
-            The prompt to provide the model. This is typically used to correct
-            spelling. Limit of 250 characters.
-          </Description>
-          <Select>
-            <SelectTrigger className="w-[180px]">
-              <SelectValue placeholder="Theme" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="light">Default</SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-      )}
+    <div className="flex items-center">
+      <Label className="flex items-center">
+        <Checkbox
+          className="mr-2"
+          checked={startupEnabled}
+          onCheckedChange={handleStartupEnabledChange}
+        />
+        <span>Start at login</span>
+      </Label>
     </div>
+  );
+}
+
+type SoundSelectProps = {
+  label: string;
+  soundEvent: string;
+  disabled: boolean;
+};
+
+function SoundSelect({
+  label,
+  soundEvent,
+  disabled = false,
+}: SoundSelectProps) {
+  const [soundSetting, setSoundSetting] = useSetting<string>(
+    soundEvent,
+    'none'
+  );
+
+  function handleChange(value: string) {
+    setSoundSetting(value);
+  }
+
+  return (
+    <>
+      <div className="justify-self-end text-sm">{label}</div>
+      <Select
+        value={soundSetting ?? 'none'}
+        onValueChange={handleChange}
+        disabled={disabled}
+      >
+        <SelectTrigger className="w-[180px] h-8">
+          <SelectValue />
+        </SelectTrigger>
+        <SelectContent>
+          <SelectItem value="none">
+            <span className="text-slate-600">None</span>
+          </SelectItem>
+          <SelectItem value="tick.mp3">Tick</SelectItem>
+        </SelectContent>
+      </Select>
+    </>
   );
 }