+++ src/windows/settings/SettingsWindow.tsx
@@ -14,10 +14,12 @@ export function Settings() {
   return (
     <div className="p-4 flex flex-col gap-8 h-full">
       <Tabs defaultValue={tab} className="" onValueChange={setTab}>
-        <TabsList>
-          <TabsTrigger value="general">General</TabsTrigger>
-          <TabsTrigger value="models">Models</TabsTrigger>
-        </TabsList>
+        <div className="flex justify-center">
+          <TabsList>
+            <TabsTrigger value="general">General</TabsTrigger>
+            <TabsTrigger value="models">Models</TabsTrigger>
+          </TabsList>
+        </div>
         <TabsContent value="general" className="p-4">
           <General />
         </TabsContent>